refactor(redirect): move into modules/
