+++ src/server/modules/redirect/RedirectController.ts
@@ -1,13 +1,13 @@
 import Express from 'express'
 import { inject, injectable } from 'inversify'
-import { gaTrackingId, logger } from '../config'
-import { NotFoundError } from '../util/error'
-import parseDomain from '../util/domain'
-import { DependencyIds, ERROR_404_PATH } from '../constants'
+import { gaTrackingId, logger } from '../../config'
+import { NotFoundError } from '../../util/error'
+import parseDomain from '../../util/domain'
+import { DependencyIds, ERROR_404_PATH } from '../../constants'
 import { AnalyticsLoggerService, RedirectService } from './services'
 import { RedirectType } from '.'
-import { EventAction, EventCategory } from '../services/analytics/types/enum'
-import { createPageViewHit } from '../services/analytics'
+import { EventAction, EventCategory } from '../../services/analytics/types/enum'
+import { createPageViewHit } from '../../services/analytics'
 
 const TRANSITION_PATH = 'transition-page.ejs'
 const GTAG_PATH = 'redirect.ejs'
